Ubdate 1-16c
In gline, it makes more sense to put the if condition into the for statement, hence you save two if statements 

The second if (i < lim - 1) is then unnecessary. (Its just make sense to check ( i < lim - 1 ) because you want to save the last place for the "\0") so thats why the author did it. The for loop would break anway if "i" hit (i = lim - 1 or i > lim -1 )